Horribly mangled NanoLog for SCIENCE!
This version of the NanoLog system breaks abstractions, has odd constructs,
and generally contains undocumented/untested code to help facillitate
instrumentation and benchmarking. All the features it adds should be
properly integrated in the future, but for now, it is committed as is
for backup purposes.

Additionaly, this branch will be force-pushed regularly, so do not use
or rely on the features here.